interpolating trajectory
