--- gym_linkage/__init__.py
@@ -1,3 +1,0 @@
-from gym.envs.registration import register
-
-register(id="Linkage-v0", entry_point="gym_linkage.envs:LinkageEnv")
