--- gym_linkage/envs/__init__.py
@@ -1,1 +1,0 @@
-from gym_linkage.envs.linkage_env import LinkageEnv
