+++ setup.py
@@ -1,14 +1,15 @@
 from setuptools import setup
 
 setup(
-    name="lifting-simulation-rl",
+    name="lifting_rl",
     version="0.0.1",
-    install_requires=[
-        "gym",
-        "sympy",
-        "scipy",
-        "keras-rl",
-        "keras==2.2.4",
-        "tensorflow==1.13.1",
-    ],
+    root='./',
+#    install_requires=[
+#        "gym",
+#        "sympy",
+#        "scipy",
+#        "keras-rl",
+#        "keras==2.2.4",
+#        "tensorflow==1.13.1",
+#    ],
 )
